# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the Kurban Hisse Kağıdı (Sacrificial Donation Management) backend.

All endpoints (except `/healthz` and `/auth/login`) require authentication via either:
- `Authorization: Bearer <signed-session-token>` (preferred for browser clients)
- `X-API-Key: <key>` (legacy, for server-to-server callers and tests)

Many request/response bodies use loose JSON schemas (`type: object, additionalProperties: true`)
because the application code carries the canonical Zod validators in
`artifacts/api-server/src/routes/*`. Tightening these schemas (and deriving them from the
server-side Zod definitions) is tracked as ongoing work.

 * OpenAPI spec version: 0.1.0
 */
export interface JsonObject {
  [key: string]: unknown;
}

export type JsonArray = JsonObject[];

export type HealthStatusStatus =
  (typeof HealthStatusStatus)[keyof typeof HealthStatusStatus];

export const HealthStatusStatus = {
  ok: "ok",
  degraded: "degraded",
} as const;

export type HealthStatusDbPool = {
  total: number;
  idle: number;
  active: number;
  waiting: number;
};

export type HealthStatusDb = {
  connected: boolean;
  latencyMs: number;
  pool: HealthStatusDbPool;
};

export interface HealthStatus {
  status: HealthStatusStatus;
  db: HealthStatusDb;
  uptime: number;
}

export interface LoginRequest {
  /** @minLength 1 */
  password: string;
}

export interface LoginResponse {
  success?: boolean;
  /** Signed session token (s1.<exp>.<sig>) */
  token: string;
  /** Unix epoch seconds when the token expires */
  expiresAt: number;
}

export interface PhotoToken {
  token: string;
  expiresAt: number;
}

export type BackupImportRequestMode =
  (typeof BackupImportRequestMode)[keyof typeof BackupImportRequestMode];

export const BackupImportRequestMode = {
  replace: "replace",
  merge: "merge",
} as const;

export interface BackupImportRequest {
  mode?: BackupImportRequestMode;
  /** If true, validate and report counts without writing */
  dryRun?: boolean;
  /** Required when mode=replace and dryRun=false */
  confirmReplace?: boolean;
  data: JsonObject;
}

export interface SimpleSuccess {
  success: boolean;
}

export type ErrorResponseDetailsItem = { [key: string]: unknown };

export type ErrorResponseHint = { [key: string]: unknown };

export interface ErrorResponse {
  /** Human-readable error message (Turkish) */
  error: string;
  detail?: string;
  details?: ErrorResponseDetailsItem[];
  issues?: string[];
  hint?: ErrorResponseHint;
}

/**
 * Operation succeeded
 */
export type SuccessResponse = SimpleSuccess;

export type ListAuditLogsParams = {
  limit?: number;
  offset?: number;
  entityType?: string;
  action?: string;
};

export type GlobalSearchParams = {
  q: string;
};