# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    API specification for the Kurban Hisse Kağıdı (Sacrificial Donation Management) backend.

    All endpoints (except `/healthz` and `/auth/login`) require authentication via either:
    - `Authorization: Bearer <signed-session-token>` (preferred for browser clients)
    - `X-API-Key: <key>` (legacy, for server-to-server callers and tests)

    Many request/response bodies use loose JSON schemas (`type: object, additionalProperties: true`)
    because the application code carries the canonical Zod validators in
    `artifacts/api-server/src/routes/*`. Tightening these schemas (and deriving them from the
    server-side Zod definitions) is tracked as ongoing work.
servers:
  - url: /api
    description: Base API path
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      security: []
      summary: Authenticate with the application password
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LoginRequest" } } }
      responses:
        "200": { description: Login successful, content: { application/json: { schema: { $ref: "#/components/schemas/LoginResponse" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
...
  /photo-token:
    get:
      operationId: getPhotoToken
      tags: [auth]
      summary: Issue a short-lived signed token for photo URLs
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PhotoToken" } } } }
        "503": { $ref: "#/components/responses/Error" }
...
  /backup/import:
    post:
      operationId: backupImport
      tags: [backup]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/BackupImportRequest" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/JsonObject" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SignedSessionToken
      description: |
        Signed session token issued by `/auth/login`. Format: `s1.<expSeconds>.<base64urlHmac>`.
        Sent as `Authorization: Bearer <token>`.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Legacy API key auth (server-to-server callers and tests). The raw API_KEY should never
        be exposed to the browser.
  responses:
...
    Error:
      description: Error response
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
...
    LoginRequest:
      type: object
      properties:
        password: { type: string, minLength: 1 }
      required: [password]
    LoginResponse:
      type: object
      properties:
        success: { type: boolean }
        token: { type: string, description: "Signed session token (s1.<exp>.<sig>)" }
        expiresAt: { type: integer, description: "Unix epoch seconds when the token expires" }
      required: [token, expiresAt]
    PhotoToken:
      type: object
      properties:
        token: { type: string }
        expiresAt: { type: integer }
      required: [token, expiresAt]
    BackupImportRequest:
      type: object
      properties:
        mode: { type: string, enum: [replace, merge] }
        dryRun: { type: boolean, description: "If true, validate and report counts without writing" }
        confirmReplace: { type: boolean, description: "Required when mode=replace and dryRun=false" }
        data:
          $ref: "#/components/schemas/JsonObject"
      required: [data]
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string, description: "Human-readable error message (Turkish)" }
        detail: { type: string }
        details: { type: array, items: { type: object, additionalProperties: true } }
        issues: { type: array, items: { type: string } }
        hint: { type: object, additionalProperties: true }
      required: [error]