# lib/api-zod/src/generated/types/photoToken.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the Kurban Hisse Kağıdı (Sacrificial Donation Management) backend.

All endpoints (except `/healthz` and `/auth/login`) require authentication via either:
- `Authorization: Bearer <signed-session-token>` (preferred for browser clients)
- `X-API-Key: <key>` (legacy, for server-to-server callers and tests)

Many request/response bodies use loose JSON schemas (`type: object, additionalProperties: true`)
because the application code carries the canonical Zod validators in
`artifacts/api-server/src/routes/*`. Tightening these schemas (and deriving them from the
server-side Zod definitions) is tracked as ongoing work.

 * OpenAPI spec version: 0.1.0
 */

export interface PhotoToken {
  token: string;
  expiresAt: number;
}